+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * Fix NRE when a particle emitter transform is missing
 * Fix exception when TestFlight is installed
+* Gracefully handle harmony patching failure (usually caused by out of date mods)
 
 ## 3.2.4 - 2025-02-05